pyproject.toml
[project]
name = "my-django-app"
version = "0.1.0"
description = "A Django application with REST API, image handling, and background jobs."
authors = [{ name = "Dennis Van Eecke" }]
requires-python = ">=3.11"
dependencies = [
  "django",
  "gunicorn",
  "django-solo",
  "psycopg[binary]",
  "djangorestframework",
  "httpx",
  "python-dateutil",
  "croniter",
  "openai",
  "pillow",
  "pi-heif",
  "pydantic",
  "psutil",
  "docker",
  "requests",
  "ics",
  "playwright",
]

[project.optional-dependencies]
dev = [
  "ruff",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
